refactor: YoutubeVideoSection 컴포넌트를 서버 컴포넌트로 변경 및 위치 이동 (#81)
- YoutubeVideoSection을 Client → Server Component로 전환하여 렌더링 최적화
- YouTube API 호출 로직을 async 함수 기반 서버 컴포넌트로 이동
- axios 및 useEffect 의존 제거하여 불필요한 클라이언트 사이드 동작 최소화
- game/detail → games/[id]/_components 디렉토리로 위치 재조정
